Settle on a consistent plugin naming scheme.
