+++ sphinxcontrib/bibtex/style/referencing/extra_year.py
@@ -1,0 +1,30 @@
+import dataclasses
+
+from typing import TYPE_CHECKING, List, Iterable
+from pybtex.style.template import field
+from sphinxcontrib.bibtex.richtext import ReferenceInfo
+from sphinxcontrib.bibtex.style.template import reference
+from . import BaseStandardReferenceStyle
+
+if TYPE_CHECKING:
+    from pybtex.richtext import BaseText
+    from pybtex.style.template import Node
+
+
+@dataclasses.dataclass(frozen=True)
+class ExtraYearReferenceStyle(BaseStandardReferenceStyle[ReferenceInfo]):
+    """Reference by year."""
+
+    def get_role_names(self) -> Iterable[str]:
+        return ['year', 'yearpar']
+
+    def get_outer_template(
+            self, role_name: str, children: List["BaseText"]) -> "Node":
+        return self.get_standard_outer_template(
+            children,
+            brackets='par' in role_name,
+            capfirst=False,
+        )
+
+    def get_inner_template(self, role_name: str) -> "Node":
+        return reference[field('year')]
