+++ sphinxcontrib/bibtex/style/referencing/label.py
@@ -4,14 +4,14 @@ from sphinxcontrib.bibtex.style.referencing import (
     ReferenceInfo, BaseStandardReferenceStyle, BaseNamesReferenceStyle,
     BaseGroupReferenceStyle
 )
-from .basic_numbers import BasicNumbersReferenceStyle
-from .basic_author import BasicAuthorReferenceStyle
 from .basic_label import BasicLabelReferenceStyle
-from .basic_year import BasicYearReferenceStyle
+from .extra_author import ExtraAuthorReferenceStyle
+from .extra_label import ExtraLabelReferenceStyle
+from .extra_year import ExtraYearReferenceStyle
 
 
 @dataclasses.dataclass(frozen=True)
-class NumbersReferenceStyle(
+class LabelReferenceStyle(
         BaseGroupReferenceStyle[ReferenceInfo],
         BaseNamesReferenceStyle[ReferenceInfo],
         BaseStandardReferenceStyle[ReferenceInfo],
@@ -19,31 +19,31 @@ class NumbersReferenceStyle(
 
     def __post_init__(self):
         self.styles.extend([
-            BasicNumbersReferenceStyle(
-                ReferenceText=self.ReferenceText,
-                name_style_plugin=self.name_style_plugin,
-                abbreviate_names=self.abbreviate_names,
-                left_bracket=self.left_bracket,
-                right_bracket=self.right_bracket,
-                outer_separators=self.outer_separators,
-                names_separators=self.names_separators,
-            ),
-            BasicAuthorReferenceStyle(
-                ReferenceText=self.ReferenceText,
-                name_style_plugin=self.name_style_plugin,
-                abbreviate_names=self.abbreviate_names,
-                left_bracket=self.left_bracket,
-                right_bracket=self.right_bracket,
-                outer_separators=self.outer_separators,
-                names_separators=self.names_separators,
-            ),
             BasicLabelReferenceStyle(
                 ReferenceText=self.ReferenceText,
+                name_style_plugin=self.name_style_plugin,
+                abbreviate_names=self.abbreviate_names,
+                left_bracket=self.left_bracket,
+                right_bracket=self.right_bracket,
+                outer_separators=self.outer_separators,
+                names_separators=self.names_separators,
+            ),
+            ExtraAuthorReferenceStyle(
+                ReferenceText=self.ReferenceText,
+                name_style_plugin=self.name_style_plugin,
+                abbreviate_names=self.abbreviate_names,
+                left_bracket=self.left_bracket,
+                right_bracket=self.right_bracket,
+                outer_separators=self.outer_separators,
+                names_separators=self.names_separators,
+            ),
+            ExtraLabelReferenceStyle(
+                ReferenceText=self.ReferenceText,
                 left_bracket=self.left_bracket,
                 right_bracket=self.right_bracket,
                 outer_separators=self.outer_separators,
             ),
-            BasicYearReferenceStyle(
+            ExtraYearReferenceStyle(
                 ReferenceText=self.ReferenceText,
                 left_bracket=self.left_bracket,
                 right_bracket=self.right_bracket,
